pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kr-cli-dominion"
version = "5.3.35"
description = "Advanced AI-Powered Security CLI - DOMINION Model"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "KaliRootCode", email = "kalirootcode@proton.me"}
]
keywords = ["cybersecurity", "hacking", "pentesting", "kali", "termux", "cli", "ai", "agent"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Android",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "groq>=0.4.0",
    "supabase>=2.0.0",
    "reportlab>=4.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
# Web search (requires compilation, may fail on Termux)
websearch = [
    "duckduckgo-search>=6.0.0",
]
# Audio features (may not work on Termux/Android)
audio = [
    "sounddevice>=0.4.6",
    "scipy>=1.11.0",
    "numpy>=1.26.0",
]
# Full PC installation
full = [
    "duckduckgo-search>=6.0.0",
    "sounddevice>=0.4.6",
    "scipy>=1.11.0",
    "numpy>=1.26.0",
]

[project.urls]
Homepage = "https://github.com/kalirootcode/KaliRootCLI"
Repository = "https://github.com/kalirootcode/KaliRootCLI"
Issues = "https://github.com/kalirootcode/KaliRootCLI/issues"

[project.scripts]
kr-clidn = "kalirootcli.main:main"
kr-cli = "kalirootcli.cli_wrapper:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["kalirootcli*"]
